+++ packages/bookings/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @voyant-travel/bookings
 
+## 0.138.9
+
 ## 0.138.8
 
 ## 0.138.7